fixes to overlay and UI
