+++ src/routeTree.gen.ts
@@ -19,8 +19,6 @@ import { Route as TemplatesNewRouteImport } from './routes/templates.new'
 import { Route as TemplatesIndexRouteImport } from './routes/templates._index'
 import { Route as TemplatesIdRouteImport } from './routes/templates.$id'
 import { Route as ProgramsIndexRouteImport } from './routes/programs._index'
-import { Route as HistoryIndexRouteImport } from './routes/history._index'
-import { Route as HistoryExerciseIdRouteImport } from './routes/history.$exerciseId'
 import { Route as ExercisesNewRouteImport } from './routes/exercises.new'
 import { Route as ExercisesIndexRouteImport } from './routes/exercises._index'
 import { Route as ExercisesIdRouteImport } from './routes/exercises.$id'
@@ -76,7 +74,6 @@ import { Route as ApiProgramCyclesIdCreate1rmTestWorkoutRouteImport } from './ro
 import { Route as ApiProgramCyclesCycleId1rmTestWorkoutRouteImport } from './routes/api/program-cycles.$cycleId.1rm-test-workout'
 import { Route as ApiExercisesIdLastWorkoutSetsRouteImport } from './routes/api/exercises.$id.last-workout-sets'
 import { Route as ApiExercisesIdLastWorkoutRouteImport } from './routes/api/exercises.$id.last-workout'
-import { Route as ApiExercisesExerciseIdHistoryRouteImport } from './routes/api/exercises.$exerciseId.history'
 import { Route as ApiWorkoutsIdExercisesReorderRouteImport } from './routes/api/workouts.$id.exercises.reorder'
 import { Route as ApiTemplatesIdExercisesReorderRouteImport } from './routes/api/templates.$id.exercises.reorder'
 import { Route as ApiTemplatesIdExercisesExerciseIdRouteImport } from './routes/api/templates.$id.exercises.$exerciseId'
@@ -130,16 +127,6 @@ const TemplatesIdRoute = TemplatesIdRouteImport.update({
 const ProgramsIndexRoute = ProgramsIndexRouteImport.update({
   id: '/programs/_index',
   path: '/programs',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const HistoryIndexRoute = HistoryIndexRouteImport.update({
-  id: '/history/_index',
-  path: '/history',
-  getParentRoute: () => rootRouteImport,
-} as any)
-const HistoryExerciseIdRoute = HistoryExerciseIdRouteImport.update({
-  id: '/history/$exerciseId',
-  path: '/history/$exerciseId',
   getParentRoute: () => rootRouteImport,
 } as any)
 const ExercisesNewRoute = ExercisesNewRouteImport.update({
@@ -428,12 +415,6 @@ const ApiExercisesIdLastWorkoutRoute =
     path: '/last-workout',
     getParentRoute: () => ApiExercisesIdRoute,
   } as any)
-const ApiExercisesExerciseIdHistoryRoute =
-  ApiExercisesExerciseIdHistoryRouteImport.update({
-    id: '/$exerciseId/history',
-    path: '/$exerciseId/history',
-    getParentRoute: () => ApiExercisesRoute,
-  } as any)
 const ApiWorkoutsIdExercisesReorderRoute =
   ApiWorkoutsIdExercisesReorderRouteImport.update({
     id: '/reorder',
@@ -480,8 +461,6 @@ export interface FileRoutesByFullPath {
   '/exercises/$id': typeof ExercisesIdRouteWithChildren
   '/exercises': typeof ExercisesIndexRoute
   '/exercises/new': typeof ExercisesNewRoute
-  '/history/$exerciseId': typeof HistoryExerciseIdRoute
-  '/history': typeof HistoryIndexRoute
   '/programs': typeof ProgramsIndexRoute
   '/templates/$id': typeof TemplatesIdRouteWithChildren
   '/templates': typeof TemplatesIndexRoute
@@ -512,7 +491,6 @@ export interface FileRoutesByFullPath {
   '/workouts/$id/edit': typeof WorkoutsIdEditRoute
   '/workouts/$id/summary': typeof WorkoutsIdSummaryRoute
   '/workouts/start/$templateId': typeof WorkoutsStartTemplateIdRoute
-  '/api/exercises/$exerciseId/history': typeof ApiExercisesExerciseIdHistoryRoute
   '/api/exercises/$id/last-workout': typeof ApiExercisesIdLastWorkoutRoute
   '/api/exercises/$id/last-workout-sets': typeof ApiExercisesIdLastWorkoutSetsRoute
   '/api/program-cycles/$cycleId/1rm-test-workout': typeof ApiProgramCyclesCycleId1rmTestWorkoutRoute
@@ -554,8 +532,6 @@ export interface FileRoutesByTo {
   '/exercises/$id': typeof ExercisesIdRouteWithChildren
   '/exercises': typeof ExercisesIndexRoute
   '/exercises/new': typeof ExercisesNewRoute
-  '/history/$exerciseId': typeof HistoryExerciseIdRoute
-  '/history': typeof HistoryIndexRoute
   '/programs': typeof ProgramsIndexRoute
   '/templates/$id': typeof TemplatesIdRouteWithChildren
   '/templates': typeof TemplatesIndexRoute
@@ -586,7 +562,6 @@ export interface FileRoutesByTo {
   '/workouts/$id/edit': typeof WorkoutsIdEditRoute
   '/workouts/$id/summary': typeof WorkoutsIdSummaryRoute
   '/workouts/start/$templateId': typeof WorkoutsStartTemplateIdRoute
-  '/api/exercises/$exerciseId/history': typeof ApiExercisesExerciseIdHistoryRoute
   '/api/exercises/$id/last-workout': typeof ApiExercisesIdLastWorkoutRoute
   '/api/exercises/$id/last-workout-sets': typeof ApiExercisesIdLastWorkoutSetsRoute
   '/api/program-cycles/$cycleId/1rm-test-workout': typeof ApiProgramCyclesCycleId1rmTestWorkoutRoute
@@ -629,8 +604,6 @@ export interface FileRoutesById {
   '/exercises/$id': typeof ExercisesIdRouteWithChildren
   '/exercises/_index': typeof ExercisesIndexRoute
   '/exercises/new': typeof ExercisesNewRoute
-  '/history/$exerciseId': typeof HistoryExerciseIdRoute
-  '/history/_index': typeof HistoryIndexRoute
   '/programs/_index': typeof ProgramsIndexRoute
   '/templates/$id': typeof TemplatesIdRouteWithChildren
   '/templates/_index': typeof TemplatesIndexRoute
@@ -661,7 +634,6 @@ export interface FileRoutesById {
   '/workouts/$id_/edit': typeof WorkoutsIdEditRoute
   '/workouts/$id_/summary': typeof WorkoutsIdSummaryRoute
   '/workouts/start/$templateId': typeof WorkoutsStartTemplateIdRoute
-  '/api/exercises/$exerciseId/history': typeof ApiExercisesExerciseIdHistoryRoute
   '/api/exercises/$id/last-workout': typeof ApiExercisesIdLastWorkoutRoute
   '/api/exercises/$id/last-workout-sets': typeof ApiExercisesIdLastWorkoutSetsRoute
   '/api/program-cycles/$cycleId/1rm-test-workout': typeof ApiProgramCyclesCycleId1rmTestWorkoutRoute
@@ -705,8 +677,6 @@ export interface FileRouteTypes {
     | '/exercises/$id'
     | '/exercises'
     | '/exercises/new'
-    | '/history/$exerciseId'
-    | '/history'
     | '/programs'
     | '/templates/$id'
     | '/templates'
@@ -737,7 +707,6 @@ export interface FileRouteTypes {
     | '/workouts/$id/edit'
     | '/workouts/$id/summary'
     | '/workouts/start/$templateId'
-    | '/api/exercises/$exerciseId/history'
     | '/api/exercises/$id/last-workout'
     | '/api/exercises/$id/last-workout-sets'
     | '/api/program-cycles/$cycleId/1rm-test-workout'
@@ -779,8 +748,6 @@ export interface FileRouteTypes {
     | '/exercises/$id'
     | '/exercises'
     | '/exercises/new'
-    | '/history/$exerciseId'
-    | '/history'
     | '/programs'
     | '/templates/$id'
     | '/templates'
@@ -811,7 +778,6 @@ export interface FileRouteTypes {
     | '/workouts/$id/edit'
     | '/workouts/$id/summary'
     | '/workouts/start/$templateId'
-    | '/api/exercises/$exerciseId/history'
     | '/api/exercises/$id/last-workout'
     | '/api/exercises/$id/last-workout-sets'
     | '/api/program-cycles/$cycleId/1rm-test-workout'
@@ -853,8 +819,6 @@ export interface FileRouteTypes {
     | '/exercises/$id'
     | '/exercises/_index'
     | '/exercises/new'
-    | '/history/$exerciseId'
-    | '/history/_index'
     | '/programs/_index'
     | '/templates/$id'
     | '/templates/_index'
@@ -885,7 +849,6 @@ export interface FileRouteTypes {
     | '/workouts/$id_/edit'
     | '/workouts/$id_/summary'
     | '/workouts/start/$templateId'
-    | '/api/exercises/$exerciseId/history'
     | '/api/exercises/$id/last-workout'
     | '/api/exercises/$id/last-workout-sets'
     | '/api/program-cycles/$cycleId/1rm-test-workout'
@@ -928,8 +891,6 @@ export interface RootRouteChildren {
   ExercisesIdRoute: typeof ExercisesIdRouteWithChildren
   ExercisesIndexRoute: typeof ExercisesIndexRoute
   ExercisesNewRoute: typeof ExercisesNewRoute
-  HistoryExerciseIdRoute: typeof HistoryExerciseIdRoute
-  HistoryIndexRoute: typeof HistoryIndexRoute
   ProgramsIndexRoute: typeof ProgramsIndexRoute
   TemplatesIdRoute: typeof TemplatesIdRouteWithChildren
   TemplatesIndexRoute: typeof TemplatesIndexRoute
@@ -1025,20 +986,6 @@ declare module '@tanstack/react-router' {
       path: '/programs'
       fullPath: '/programs'
       preLoaderRoute: typeof ProgramsIndexRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/history/_index': {
-      id: '/history/_index'
-      path: '/history'
-      fullPath: '/history'
-      preLoaderRoute: typeof HistoryIndexRouteImport
-      parentRoute: typeof rootRouteImport
-    }
-    '/history/$exerciseId': {
-      id: '/history/$exerciseId'
-      path: '/history/$exerciseId'
-      fullPath: '/history/$exerciseId'
-      preLoaderRoute: typeof HistoryExerciseIdRouteImport
       parentRoute: typeof rootRouteImport
     }
     '/exercises/new': {
@@ -1426,13 +1373,6 @@ declare module '@tanstack/react-router' {
       preLoaderRoute: typeof ApiExercisesIdLastWorkoutRouteImport
       parentRoute: typeof ApiExercisesIdRoute
     }
-    '/api/exercises/$exerciseId/history': {
-      id: '/api/exercises/$exerciseId/history'
-      path: '/$exerciseId/history'
-      fullPath: '/api/exercises/$exerciseId/history'
-      preLoaderRoute: typeof ApiExercisesExerciseIdHistoryRouteImport
-      parentRoute: typeof ApiExercisesRoute
-    }
     '/api/workouts/$id/exercises/reorder': {
       id: '/api/workouts/$id/exercises/reorder'
       path: '/reorder'
@@ -1481,13 +1421,11 @@ const ApiExercisesIdRouteWithChildren = ApiExercisesIdRoute._addFileChildren(
 interface ApiExercisesRouteChildren {
   ApiExercisesIdRoute: typeof ApiExercisesIdRouteWithChildren
   ApiExercisesCopyFromLibraryRoute: typeof ApiExercisesCopyFromLibraryRoute
-  ApiExercisesExerciseIdHistoryRoute: typeof ApiExercisesExerciseIdHistoryRoute
 }
 
 const ApiExercisesRouteChildren: ApiExercisesRouteChildren = {
   ApiExercisesIdRoute: ApiExercisesIdRouteWithChildren,
   ApiExercisesCopyFromLibraryRoute: ApiExercisesCopyFromLibraryRoute,
-  ApiExercisesExerciseIdHistoryRoute: ApiExercisesExerciseIdHistoryRoute,
 }
 
 const ApiExercisesRouteWithChildren = ApiExercisesRoute._addFileChildren(
@@ -1687,8 +1625,6 @@ const rootRouteChildren: RootRouteChildren = {
   ExercisesIdRoute: ExercisesIdRouteWithChildren,
   ExercisesIndexRoute: ExercisesIndexRoute,
   ExercisesNewRoute: ExercisesNewRoute,
-  HistoryExerciseIdRoute: HistoryExerciseIdRoute,
-  HistoryIndexRoute: HistoryIndexRoute,
   ProgramsIndexRoute: ProgramsIndexRoute,
   TemplatesIdRoute: TemplatesIdRouteWithChildren,
   TemplatesIndexRoute: TemplatesIndexRoute,
