+++ vite.config.ts
@@ -1,5 +1,4 @@
 import { defineConfig } from 'vite'
-import { devtools } from '@tanstack/devtools-vite'
 import { tanstackStart } from '@tanstack/react-start/plugin/vite'
 import viteReact from '@vitejs/plugin-react'
 import viteTsConfigPaths from 'vite-tsconfig-paths'
@@ -18,7 +17,6 @@ const config = defineConfig({
     noExternal: true,
   },
   plugins: [
-    devtools(),
     cloudflare({
       viteEnvironment: { name: 'ssr' },
     }),
